requirements.txt
flask==3.1.0
google-genai==1.14.0
google-cloud-aiplatform==1.78.0
google-cloud-translate==3.18.0
google-cloud-logging==3.11.3
google-cloud-firestore==2.19.0
google-cloud-secret-manager==2.21.1
google-cloud-storage==2.19.0
gunicorn==23.0.0
python-dotenv==1.1.0
